requirements.txt
numpy>=1.24.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
openai>=1.0.0
